PR from Xiao with changes for mac

+++ src/Font.cpp
@@ -17,6 +17,7 @@ Font::Font(int screenWidth, int screenHeight, int _fontHeight, std::string font)
 
 	//Sets the size (fontHeight pixels high, width dynamically calculated)
 	FT_Set_Pixel_Sizes(face, 0, fontHeight);
+	/*
 
 	GLCall(glPixelStorei(GL_UNPACK_ALIGNMENT, 1)); // Disable byte-alignment restriction
 
@@ -53,7 +54,7 @@ Font::Font(int screenWidth, int screenHeight, int _fontHeight, std::string font)
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x
+			static_cast<GLuint>(face->glyph->advance.x)
 		};
 		characters.insert(std::pair<GLchar, Character>(c, character));
 	}
@@ -75,6 +76,7 @@ Font::Font(int screenWidth, int screenHeight, int _fontHeight, std::string font)
 	GLCall(glVertexAttribPointer(0, 4, GL_FLOAT, GL_FALSE, 4 * sizeof(GLfloat), 0));
 	GLCall(glBindBuffer(GL_ARRAY_BUFFER, 0));
 	GLCall(glBindVertexArray(0));
+	*/
 }
 
 void Font::UpdateScreenSize(int screenWidth, int screenHeight) {
